backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
pydantic==2.10.3
python-dotenv==1.0.1
uvicorn==0.32.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.55.0
pydantic-settings==2.7.0
alembic==1.14.0
pytest==8.3.3
httpx==0.27.2